Allow user to specify libraries to be copied to bundle.
Add a copy user libs step so that the final bundle can be totally
dependency free. To do this, first you have to copy all the libs to the
bundle before compiling. Then, when compiling, you change the linker to
only look for libs inside the bundle, so that any compiled ccalls
pointing to the bundle will succeed during compilation. Finally, when the binary is
executed, it sees those libs b/c we've add them to the @rpath.

Since the SDL libraries were being provided as absolute paths to ccall,
and since those paths get compiled in, I also had to add an ENVIRONMENT
VARIABLE set only during compilation which tells the julia code that a
mac bundle is being compiled. This lets the julia code overwrite the SDL
libraries' locations only during mac bundle compilation. The user julia
code should use relative paths for the dlls, since they need to work
both during compilation (when LD_LIBRARY_PATH will be an absolute path
to the Libraries dir), and during runtime (when the @rpath will be
@executable_path/../Libraries).

I *THINK* this is a finally a totally portable, self-contained .app bundle!!!!
(It's 273MB without assets but with SDL libs.)