Maybe fixed quit hang: close callback
Before calling `SDL_Quit()`, delete the callback. It seems like this
fixed the game hanging on quit. Hard to tell though since it only
happens sometimes.

I also tried just skipping all the SDL quit functions, and that also
seemed to fix the hang, and it quit the fastest. But I've read online
that on some systems (maybe linux) it can cause problems if you've like
gone fullscreen and then when you leave fullscreen it doesn't reset the
video mode correctly... So i'll stick with this for now.

Either way, it can't *hurt* to clean up the callback before quitting,
except that it's just a waste of the user's time to wait for it.