Fix Quit: allow quitting when app lacks focus.
Add check to `while window_paused[]; poll; end` loop to also keep an eye
out for QUIT events. Also throw a QuitException if a QUIT event occurs
while the window is paused. Before, these would be ignored, the same as
any of the other events.

Bumped version number to v.1.0.2